+++ testing/macro-tests/tests/compile.rs
@@ -1,5 +1,6 @@
 #[test]
-fn compile_pass() {
+fn compile_ui() {
     let t = trybuild::TestCases::new();
+    t.compile_fail("tests/ui/*.rs");
     t.pass("tests/pass/*.rs");
 }